filesystem: Move IStream implementations out of Common
