+++ src/3D/Sky.cpp
@@ -15,9 +15,10 @@
 
 #include "3D/L3DMesh.h"
 #include "Common/Bitmap16B.h"
-#include "Common/FileSystem.h"
 #include "Common/StringUtils.h"
+#include "FileSystem/FileSystem.h"
 
+using namespace openblack::filesystem;
 using namespace openblack::graphics;
 
 namespace openblack
