+++ src/Debug/Gui.cpp
@@ -43,11 +43,11 @@
 
 #include <3D/Camera.h>
 #include <3D/Sky.h>
-#include <Common/FileSystem.h>
 #include <ECS/Components/LivingAction.h>
 #include <ECS/Components/Transform.h>
 #include <ECS/Components/Villager.h>
 #include <ECS/Registry.h>
+#include <FileSystem/FileSystem.h>
 #include <Game.h>
 #include <GameWindow.h>
 #include <Locator.h>
