+++ src/Graphics/ShaderProgram.cpp
@@ -11,7 +11,7 @@
 
 #include <spdlog/spdlog.h>
 
-#include "Common/FileSystem.h"
+#include "FileSystem/FileSystem.h"
 #include "Game.h"
 #include "Texture2D.h"
