+++ src/Parsers/InfoFile.cpp
@@ -12,7 +12,7 @@
 #include <PackFile.h>
 #include <spdlog/spdlog.h>
 
-#include "Common/FileSystem.h"
+#include "FileSystem/FileSystem.h"
 #include "Game.h"
 
 namespace openblack
