+++ src/Serializer/FotFile.cpp
@@ -12,13 +12,14 @@
 #include <spdlog/spdlog.h>
 
 #include "3D/LandIsland.h"
-#include "Common/FileSystem.h"
 #include "ECS/Components/Footpath.h"
 #include "ECS/Registry.h"
+#include "FileSystem/FileSystem.h"
 #include "Game.h"
 #include "GameThingSerializer.h"
 
 using namespace openblack;
+using namespace openblack::filesystem;
 
 FotFile::FotFile(Game& game)
     : _game(game)
